ref(server-utils): Fold cache handling into `redisIntegration`
