+++ packages/node/src/integrations/tracing/index.ts
@@ -36,8 +36,8 @@ export function getAutoPerformanceIntegrations(): Integration[] {
     mongooseIntegration(),
     mysqlIntegration(),
     mysql2Integration(),
-    // Redis keeps the node wrapper: it wires the cache `responseHook` into the channel subscribers
-    // and covers all redis client versions (native diagnostics_channel + orchestrion fallbacks).
+    // Redis keeps the node wrapper: it composes the ioredis + redis channel integrations into one
+    // integration covering all client versions (native diagnostics_channel + orchestrion fallbacks).
     redisIntegration(),
     postgresIntegration(),
     prismaIntegration(),
