+++ packages/server-utils/src/orchestrion/index.ts
@@ -66,7 +66,6 @@ export {
   firebaseIntegration,
 };
 export type { KoaIntegrationOptions } from '../integrations/koa';
-export type { IORedisChannelIntegrationOptions, IORedisResponseHook } from '../integrations/ioredis';
 export type { PostgresJsIntegrationOptions } from '../integrations/postgres-js';
 export { redisIntegration } from '../integrations/redis';
 export type { InstrumentationConfig, CustomTransform } from './apmTypes';
@@ -82,9 +81,9 @@ export type * from '../integrations/graphql/graphql-types';
  * Single source of truth: add a new channel integration here and every consumer that spreads this map
  * into its default integrations picks it up automatically, so there's no separate list to keep in sync.
  *
- * NOTE: `ioredisChannelIntegration` and `redisIntegration` are intentionally NOT here. They
- * only partially replace the composite OTel `Redis` integration and need the node SDK's redis cache
- * `responseHook` (which can't live in `server-utils`), so `@sentry/node` wires them up separately.
+ * NOTE: `ioredisChannelIntegration` and `redisIntegration` are intentionally NOT here. They only
+ * partially replace the composite OTel `Redis` integration, so `@sentry/node` composes them into a
+ * single `Redis` integration (gated against the OTel one) and wires them up separately.
  *
  * Framework SDKs that own their own channel listener (e.g. `@sentry/nestjs`'s `Nest`) are NOT here
  * either: their transform config is still in `SENTRY_INSTRUMENTATIONS`, but the listener lives in
